backend/requirements.txt
anthropic>=0.25.0
chromadb>=0.5.0
pandas>=2.0.0
openpyxl>=3.1.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
folium>=0.16.0
python-dotenv>=1.0.0
sentence-transformers>=2.7.0
pydantic>=2.0.0
numpy>=1.26.0
mlflow>=2.13.0
